# backend/openapi/schema.yaml
openapi: 3.0.3
info:
  title: Project Meridian API
  version: v1
  description: Product lifecycle management platform API.
paths:
...
  /api/v1/pilot/batches/{batch_public_id}/feedback:
    get:
      operationId: pilot_feedback_list
      parameters:
      - in: path
        name: batch_public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - pilot
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotFeedbackList'
          description: ''
    post:
      operationId: pilot_feedback_create
      parameters:
      - in: path
        name: batch_public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - pilot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotFeedbackCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PilotFeedbackCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PilotFeedbackCreateRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotFeedbackCreateResponse'
          description: ''
...
  /api/v1/pilot/feedback/{public_id}/assign:
    post:
      operationId: pilot_feedback_assign
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - pilot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotFeedbackAssignRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PilotFeedbackAssignRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PilotFeedbackAssignRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotFeedbackAssignResponse'
          description: ''
  /api/v1/pilot/feedback/{public_id}/close:
    post:
      operationId: pilot_feedback_close
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - pilot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotFeedbackCloseRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PilotFeedbackCloseRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PilotFeedbackCloseRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotFeedbackCloseResponse'
          description: ''
  /api/v1/pilot/feedback/{public_id}/handle:
    post:
      operationId: pilot_feedback_handle
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - pilot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotFeedbackHandleRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PilotFeedbackHandleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PilotFeedbackHandleRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotFeedbackHandleResponse'
          description: ''
  /api/v1/pilot/feedback/{public_id}/retest:
    post:
      operationId: pilot_feedback_retest
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - pilot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotFeedbackRetestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PilotFeedbackRetestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PilotFeedbackRetestRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotFeedbackRetestResponse'
          description: ''
  /api/v1/pilot/feedback/{public_id}/submit-retest:
    post:
      operationId: pilot_feedback_submit_retest
      parameters:
      - in: path
        name: public_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - pilot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PilotFeedbackSubmitRetestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PilotFeedbackSubmitRetestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PilotFeedbackSubmitRetestRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PilotFeedbackSubmitRetestResponse'
          description: ''
...
components:
  schemas:
...
    PilotFeedbackAssignRequest:
      type: object
      properties:
        severity:
          type: string
        assignee_public_id:
          type: string
          format: uuid
        expected_version:
          type: integer
      required:
      - assignee_public_id
      - severity
    PilotFeedbackAssignResponse:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        batch_public_id:
          type: string
          format: uuid
        title:
          type: string
        reproduction_summary:
          type: string
        severity:
          type: string
        status:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
          nullable: true
        assignee_public_id:
          type: string
          format: uuid
          nullable: true
        target_version:
          type: string
        workaround:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
          nullable: true
        acceptance_note:
          type: string
        close_reason:
          type: string
        retest_result:
          type: string
        version_no:
          type: integer
      required:
      - acceptance_note
      - accepted_by_public_id
      - assignee_public_id
      - batch_public_id
      - close_reason
      - evidence_document_version_public_id
      - external_key
      - public_id
      - reproduction_summary
      - retest_result
      - severity
      - status
      - target_version
      - title
      - version_no
      - workaround
    PilotFeedbackCloseRequest:
      type: object
      properties:
        reject:
          type: boolean
        close_reason:
          type: string
        workaround:
          type: string
        target_version:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
        acceptance_note:
          type: string
        expected_version:
          type: integer
    PilotFeedbackCloseResponse:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        batch_public_id:
          type: string
          format: uuid
        title:
          type: string
        reproduction_summary:
          type: string
        severity:
          type: string
        status:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
          nullable: true
        assignee_public_id:
          type: string
          format: uuid
          nullable: true
        target_version:
          type: string
        workaround:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
          nullable: true
        acceptance_note:
          type: string
        close_reason:
          type: string
        retest_result:
          type: string
        version_no:
          type: integer
      required:
      - acceptance_note
      - accepted_by_public_id
      - assignee_public_id
      - batch_public_id
      - close_reason
      - evidence_document_version_public_id
      - external_key
      - public_id
      - reproduction_summary
      - retest_result
      - severity
      - status
      - target_version
      - title
      - version_no
      - workaround
    PilotFeedbackCreateRequest:
      type: object
      properties:
        title:
          type: string
        reproduction_summary:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
      required:
      - reproduction_summary
      - title
    PilotFeedbackCreateResponse:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        batch_public_id:
          type: string
          format: uuid
        title:
          type: string
        reproduction_summary:
          type: string
        severity:
          type: string
        status:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
          nullable: true
        assignee_public_id:
          type: string
          format: uuid
          nullable: true
        target_version:
          type: string
        workaround:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
          nullable: true
        acceptance_note:
          type: string
        close_reason:
          type: string
        retest_result:
          type: string
        version_no:
          type: integer
      required:
      - acceptance_note
      - accepted_by_public_id
      - assignee_public_id
      - batch_public_id
      - close_reason
      - evidence_document_version_public_id
      - external_key
      - public_id
      - reproduction_summary
      - retest_result
      - severity
      - status
      - target_version
      - title
      - version_no
      - workaround
    PilotFeedbackHandleRequest:
      type: object
      properties:
        expected_version:
          type: integer
    PilotFeedbackHandleResponse:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        batch_public_id:
          type: string
          format: uuid
        title:
          type: string
        reproduction_summary:
          type: string
        severity:
          type: string
        status:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
          nullable: true
        assignee_public_id:
          type: string
          format: uuid
          nullable: true
        target_version:
          type: string
        workaround:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
          nullable: true
        acceptance_note:
          type: string
        close_reason:
          type: string
        retest_result:
          type: string
        version_no:
          type: integer
      required:
      - acceptance_note
      - accepted_by_public_id
      - assignee_public_id
      - batch_public_id
      - close_reason
      - evidence_document_version_public_id
      - external_key
      - public_id
      - reproduction_summary
      - retest_result
      - severity
      - status
      - target_version
      - title
      - version_no
      - workaround
    PilotFeedbackItem:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        batch_public_id:
          type: string
          format: uuid
        title:
          type: string
        reproduction_summary:
          type: string
        severity:
          type: string
        status:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
          nullable: true
        assignee_public_id:
          type: string
          format: uuid
          nullable: true
        target_version:
          type: string
        workaround:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
          nullable: true
        acceptance_note:
          type: string
        close_reason:
          type: string
        retest_result:
          type: string
        version_no:
          type: integer
      required:
      - acceptance_note
      - accepted_by_public_id
      - assignee_public_id
      - batch_public_id
      - close_reason
      - evidence_document_version_public_id
      - external_key
      - public_id
      - reproduction_summary
      - retest_result
      - severity
      - status
      - target_version
      - title
      - version_no
      - workaround
    PilotFeedbackList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PilotFeedbackItem'
      required:
      - items
    PilotFeedbackRetestRequest:
      type: object
      properties:
        passed:
          type: boolean
        expected_version:
          type: integer
      required:
      - passed
    PilotFeedbackRetestResponse:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        batch_public_id:
          type: string
          format: uuid
        title:
          type: string
        reproduction_summary:
          type: string
        severity:
          type: string
        status:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
          nullable: true
        assignee_public_id:
          type: string
          format: uuid
          nullable: true
        target_version:
          type: string
        workaround:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
          nullable: true
        acceptance_note:
          type: string
        close_reason:
          type: string
        retest_result:
          type: string
        version_no:
          type: integer
      required:
      - acceptance_note
      - accepted_by_public_id
      - assignee_public_id
      - batch_public_id
      - close_reason
      - evidence_document_version_public_id
      - external_key
      - public_id
      - reproduction_summary
      - retest_result
      - severity
      - status
      - target_version
      - title
      - version_no
      - workaround
    PilotFeedbackSubmitRetestRequest:
      type: object
      properties:
        target_version:
          type: string
        expected_version:
          type: integer
    PilotFeedbackSubmitRetestResponse:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        batch_public_id:
          type: string
          format: uuid
        title:
          type: string
        reproduction_summary:
          type: string
        severity:
          type: string
        status:
          type: string
        external_key:
          type: string
        evidence_document_version_public_id:
          type: string
          format: uuid
          nullable: true
        assignee_public_id:
          type: string
          format: uuid
          nullable: true
        target_version:
          type: string
        workaround:
          type: string
        accepted_by_public_id:
          type: string
          format: uuid
          nullable: true
        acceptance_note:
          type: string
        close_reason:
          type: string
        retest_result:
          type: string
        version_no:
          type: integer
      required:
      - acceptance_note
      - accepted_by_public_id
      - assignee_public_id
      - batch_public_id
      - close_reason
      - evidence_document_version_public_id
      - external_key
      - public_id
      - reproduction_summary
      - retest_result
      - severity
      - status
      - target_version
      - title
      - version_no
      - workaround
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid